Fixed LSTM model
I have to check the way the loss is calculated again later, but fort now the model trains and tests correctly as well. I will have to tweak and clean up the code, as well as train it based on all the 30 reno .csv files in the future too. But for now, the main issues that were present before are fixed.